fix(cargo): format
