+++ src/http/mod.rs
@@ -1,2 +1,6 @@
+pub mod schema;
+
+#[cfg(feature = "client")]
 pub mod client;
+#[cfg(feature = "server")]
 pub mod server;
