+++ src/lib.rs
@@ -2,5 +2,5 @@
 
 pub mod app;
 pub mod data;
-pub mod timestamp;
 pub mod http;
+pub mod timestamp;
